🚚 Rename: Content, Sidebar 컴포넌트를 Layout 폴더에서 분리
