+++ src/components/layout/index.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 
 import IndexPage from './IndexPage';
-import Sidebar from './Sidebar';
+import Sidebar from '../Sidebar';
 
 export default function Layout({ children }: { children: React.ReactNode }) {
   return (
